Added curved toolbar, applied it to several fragments
