+++ app/src/main/java/com/bangkit/cemil/authentication/ForgotPasswordFragment.kt
@@ -5,16 +5,22 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
 import com.bangkit.cemil.R
+import com.bangkit.cemil.databinding.FragmentForgotPasswordBinding
+import androidx.appcompat.widget.Toolbar
 
 class ForgotPasswordFragment : Fragment() {
+
+    private lateinit var binding : FragmentForgotPasswordBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
-        return inflater.inflate(R.layout.fragment_forgot_password, container, false)
+        binding = FragmentForgotPasswordBinding.inflate(inflater, container, false)
+        (activity as AppCompatActivity).setSupportActionBar(binding.materialToolbar)
+        return binding.root
     }
 
 }
