+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -7,6 +7,18 @@
     android:layout_gravity="center"
     tools:context=".authentication.ForgotPasswordFragment">
 
+
+    <com.google.android.material.appbar.MaterialToolbar
+        android:id="@+id/materialToolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@drawable/curved_toolbar"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="?attr/actionBarTheme"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
     <TextView
         android:id="@+id/tvForgotYourPassword"
         android:layout_width="wrap_content"
@@ -89,4 +101,6 @@
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/etForgotEmail" />
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
