+++ app/src/main/res/layout/fragment_third_landing.xml
@@ -28,5 +28,4 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/imageLandingScreen" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
