+++ app/src/main/res/navigation/main_navigation.xml
@@ -7,22 +7,22 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.bangkit.cemil.HomeFragment"
+        android:name="com.bangkit.cemil.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.bangkit.cemil.SearchFragment"
+        android:name="com.bangkit.cemil.home.SearchFragment"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/historyFragment"
-        android:name="com.bangkit.cemil.HistoryFragment"
+        android:name="com.bangkit.cemil.home.HistoryFragment"
         android:label="fragment_history"
         tools:layout="@layout/fragment_history" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.bangkit.cemil.ProfileFragment"
+        android:name="com.bangkit.cemil.home.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
